README and add api module (#4)
