+++ src/api/mod.rs
@@ -1,0 +1,21 @@
+use serde_derive::Serialize;
+
+mod contracts;
+mod external_wallets;
+mod internal_wallets;
+mod staking;
+mod transactions;
+mod vaults;
+mod wallet_connect;
+
+#[derive(Debug, Serialize)]
+struct WalletCreate {
+  name: String,
+}
+
+#[derive(Debug, Serialize)]
+#[allow(dead_code)]
+pub struct WalletCreateAsset {
+  pub address: String,
+  pub tag: String,
+}
